==> 05-Ajustar_pendientes.Rmd <==
# Ajuste lineal para obtener pendientes
Para obtener las tasas (la pendiente) usamos un ajuste lineal.

A parte del la pendiente, también nos interesa otra información como la R^2^, el intervalo de confianza de la pendiente y si es significativa o no. Para ello, vamos a crear una función que ajuste una recta a nuestros datos y que además nos guarde toda la información que queremos en un objeto.
```{r}
#Una función para obtener la pendiente, la R y el intervalo de confianza
fit_rates <- function(data){
  lm <- lm(CO2_ppm~Seconds, data = data)
  cor <- cor.test(formula = ~Seconds +CO2_ppm, data = data)
  ajustes <- data.frame(R2 = as.numeric(cor$estimate^2), pvalue = cor$p.value, slope = as.numeric(lm$coefficients[2]), Intercept = as.numeric(lm$coefficients[1]), n = nrow(data), Conf_2.5 = confint(lm)[2,1], Conf_97.5 = confint(lm)[2,2])
  return(ajustes)
}
```

Una vez creada la función, podemos aplicarla a nuestros datos:

```{r}
Results <- Reduced_data %>% group_by(Plot) %>% group_modify(~fit_rates(data=.x))
```

Este es el objeto que hemos generado:

```{r, include=FALSE}
knitr::kable(Results, booktabs = T, format = "html") %>%
  kableExtra::kable_styling() %>%
  kableExtra::scroll_box(width = "100%", height = "500px")
```


Vamos a echar un vistazo a los ajustes que acabamos de realizar.

```{r}
#Now we can have a look into the plots with the fits
ggplot(Reduced_data, aes(x = DT, y = CO2_ppm))+
  geom_point()+
  facet_wrap(~Plot, scales = 'free')+
  geom_abline(data = Results, aes(intercept = Intercept, slope = slope), color = "red")+
  geom_text(data = Results, aes(x= 30, y = 410, label = paste("R2 =", round(R2, 2))), size = 2)
```

Ahora vamos a guardar la información relativa a los ajsutes:

```{r}
write_csv(Results, "Resultados_ajustes.csv")
```
