==> 06-Calcular_tasas.Rmd <==
# Calcular la tasa de emisión o absorción

## ¿Pero no tenemos ya la pendiente?

Sí, en el apartado anterior hemos ajustado una ecuación lineal a nuestros datos y hemos obtenido una pendiente, con su intervalo de confianza y la R^2. Y, en efecto, esa pendiente también es una tasa. Sin embargo, la pendiente que hemos calculado nos da una tasa en ppm·segundos^-1^. Lo que a nosotros nos interesa es expresar la tasa en unas unidadeds comparable con otros trabajos, artículos, etc. Estos, generalmente, expresan la tasa de emisión de CO~2~ en mmol·m^-2^·día^-1^ o mg·m^-2^·día^-1^ [p. ej. @Montes-Perez2022;@Montes-Perez2022b].

## ppm a mmol

Para esto vamos a basarnos en la conocida ley de los gases ideales que enuncia los siguiente:

\begin{equation}
PV = nRT (\#eq:GasesIdeales)
\end{equation}

Donde, *P* es la presión que ejerce un gas, *V* es el volumen que ocupa dicho gas, *n* es el número de moles, *R* es la constante de los gases ideales y *T* es la temperatura expresada en grados Kelvin.
La unidad de concentración que nos da el IRGA es ppm, partes por millón, en el caso de que la medida sea de una muestra de gas, esta indica el número de moles de nuestra sustancia (CO~2~) por millón de moles de aire, o lo que es lo mismo μmol/mol. Si fuera agua,
serı́an los gramos de sustancia por millón de gramos del total (μg/g).

Vamos a definir unos valores para este ejemplo.
```{r}
#definimos las condiciones
V_camara <- 1.171 #dm³
S <- 0.0078 #m²

Patm = 0.9905749 #atm #1003.7 #mBa
Temp = 290.7 #Kelvin #17.7 #ºC

R = 0.08205 #dm3·atm·mol-1·K-1

Slope = 0.5 #ppm/sec
```
Imaginamos que tenemos una pendiente de `r Slope` ppm/segundo. En primer lugar, para calcular el flujo debemos calcular cuantos moles de $CO_2$ se han producido por segundo en nuestra cámara. Nuestra pendiente ahora mismo son ppm/seg, o lo que es lo mismo $\mu$moles de CO~2~/ mol de aire / segundo. Por lo tanto, vamos a convertir los ppm ($\mu$moles de CO~2~/ mol de aire) en $\mu$moles de CO~2~. Tenemos una concentración y necesitamos una cantidad. Para ello, nos servimos de la ecuación \@ref(eq:GasesIdeales) y de los datos que nos aporta el IRGA. Convertimos la presión atmosférica de mBa a atm (1 atm = 1013.25 mBa) y la temperatura ambiente, de 17.7 $^{\circ} C$, hay que convertirla en grados Kelvin. Por lo tanto, tenemos 0.9906 atm, una temperaturade 290.7 K y sabemos que 0.5 ppm son 0.0000005 moles de $CO_2$ en 1 mol de aire. 


¿Cuántos moles de aire tengo en mi cámara?

Usando la ecuación \@ref(eq:GasesIdeales):

\begin{equation}
n_{aire} = \frac{1.019 ~atm \cdot 1.171 ~dm^{3}}{0.08205 ~ dm^3 \cdot atm \cdot mol^{-1} \cdot K^{-1}  \cdot 290.7 ~K} (\#eq:Molaire)
\end{equation}

```{r}
#Calculamos los moles de aire
n_aire <- (Patm * V_camara) / (R * Temp)
n_aire
```

Ahora con una simple regla de tres podemos saber cuantos moles de CO~2~ se producen en mi cámara por segundo. 

\begin{equation}
\frac{0.5 ~\mu mol~ CO_2}{1 ~mol~de~aire} = \frac{x ~\mu mol~ CO_2}{`r round(n_aire,3)` ~mol~de~aire} (\#eq:umolCO2)
\end{equation}

```{r}
umolCO2 <- Slope*n_aire #umol de CO2
umolCO2
```

Pues ya sabemos los $\mu$mol que se producen por segundo en mi cámara, `r umolCO2` $\mu$mol·s^-1^. 

Sabiendo la superficie de nuestra cámara y haciendo algunos cambios de unidades podemos calcular la tasa en mmol·m^-2^·día^-1^ o mg·m^-2^·día^-1^.

```{r}
tasa_mmol <- umolCO2/1000*86400/S
tasa_mmol
```

```{r}
tasa_mg <- umolCO2/1000*86400/S*44
tasa_mg
```

¡Pues listo! Ya hemos calculado la tasa de emisión y/o absorción de nuestro sedimento, suelo o masa de agua.
